TankSim/pyproject.toml
[project]
name = "TankSim"
version = "0.1.0"
authors = [ {name = "Silas Henrique", email = "silash35@gmail.com"}, ]
requires-python = ">=3.12, <3.13"

dependencies = [
    "casadi>=3.6.7",
    "matplotlib>=3.9.2",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "pyserial>=3.5",
    "PySide6>=6.8.0.2",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.4",
]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.format]
docstring-code-format = true
